Refactor: creating package util, creating exceptions classes in exceptions package and changing Error classes to util package.
