+++ src/main/java/com/jogamais/ufcg/exceptions/CourtException.java
@@ -1,0 +1,6 @@
+package com.jogamais.ufcg.exceptions;
+
+public class CourtException extends Exception {
+
+    private static final Long serialVersionUID = 1L;
+}
